Migrate projects to Stimulus
