+++ app/views/dashboard/show.html.haml
@@ -2,6 +2,7 @@
 %p= link_to "Artsy Viewer", artsy_viewer_path
 %p= link_to "Fairing Direball", "faring_direball"
 %p= link_to "Reading List", reading_list_path(year: Time.now.year)
+%p= link_to "Project List", projects_path
 %p= link_to "Root", root_path
 %p= link_to "Style Pages", article_styles_path
