+++ app/views/projects/index.html.haml
@@ -1,22 +1,20 @@
-#wrapper.projects
-  %h3 Today is a great day to make some updates!
+%h3 Today is a great day to make some updates!
 
-  %main#root
-    %div
-      %aside
-        %menu
-          %a(href="#") Create Project
-        %section
-          = form_for project do |f|
-            = f.text_field :name
-            = f.button "Create"
-      - if projects.any?
-        %ul
-          - projects.each do |project|
-            %li{"data-project-id" => project.id}
-              %span.name= project.name
-              %span.touched_at= project.touched_at
-      - else
-        %p No projects - create one!
+%div
+  %aside
+    %menu
+      %a(href="#") Create Project
+    %section
+      = form_for project do |f|
+        = f.text_field :name
+        = f.button "Create"
+  - if projects.any?
+    %ul(data-controller="projects")
+      - projects.each do |project|
+        %li{"data-project-id" => project.id}
+          %span.name= project.name
+          %span.touched_at= project.touched_at
+  - else
+    %p No projects - create one!
 
-  %h3 Try hard and believe in yourself!
+%h3 Try hard and believe in yourself!
