+++ spec/system/user_updates_project_spec.rb
@@ -4,7 +4,7 @@ describe "User updates project", js: true do
   context "signed in as admin" do
     include_context "session password matches"
 
-    xscenario "project is updated" do
+    scenario "project is updated" do
       FactoryBot.create :project, name: "Older"
       project = FactoryBot.create :project
       FactoryBot.create :project, name: "Newer"
